TimeStatを切り出し
ついでにlapCountが変わったときだけ更新するようにする
